[KGEN] Implement `pop.variant.discr_gep` operation (#44286)
This operation takes a pointer to a variant and returns a pointer to the
discriminant field using the appropriate dtype. This will simplify the
implementation of variant and allow destructuring variants
post-elaboration.

MODULAR_ORIG_COMMIT_REV_ID: efb948ab0807c29485c5ca7cec37c6039d81ab6c

+++ KGEN/lib/KGENDialect/KGENTypes.cpp
@@ -1433,7 +1433,7 @@ static std::optional<int64_t> computeVariantContentSize(VariantType type,
 /// discriminator field is the smallest integer type whose maximum value is
 /// greater than the number of possible subtypes, but which is at least `i1`.
 /// The size is rounded to the nearest integer type with a power of 2 bytewidth.
-int64_t VariantType::getDiscrSizeInBits() const {
+size_t VariantType::getDiscrSizeInBits() const {
   // Compute the smallest iN where N > 0 that fits the count.
   uint64_t bits = std::max(1u, llvm::Log2_32_Ceil(getNumTypes()));
   // Now ceildiv this to the nearest byte multiple.
